requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
boto3==1.34.83
botocore==1.34.83
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
ecdsa==0.19.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
jmespath==1.0.1
moviepy==1.0.3
numpy==1.26.4
openai==1.17.0
passlib==1.7.4
pillow==10.3.0
proglog==0.1.10
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
setuptools==69.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
